% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{annotate_last_date}
\alias{annotate_last_date}
\title{Add last date of data}
\usage{
annotate_last_date(
  font_family = getOption("fedplot.font_family"),
  font_size = 7,
  color = "black",
  nudge_x = 0,
  nudge_y = 0,
  repel = FALSE,
  box.padding = 0.25,
  point.padding = 1e-06,
  min.segment.length = 0.5,
  arrow = NULL,
  force = 1,
  force_pull = 1,
  max.time = 0.5,
  max.iter = 10000,
  max.overlaps = getOption("ggrepel.max.overlaps", default = 10),
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  direction = c("both", "y", "x"),
  seed = 1234,
  verbose = FALSE,
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  text_aes = NULL,
  ...
)
}
\arguments{
\item{font_family}{Font family. If empty, it will default to Franklin Gothic Condensed (unless the "fedplot.font_family" option is overwritten).}

\item{font_size}{Font size. If empty, will default to size 7 (unless the "fedplot.font_size" option is overwritten).}

\item{color}{Color of the text. Defaults to \code{black}.}

\item{nudge_x}{Manually nudge the x-axis of the text. Alternative to the \code{repel} option.}

\item{nudge_y}{Manually nudge the y-axis of the text. Alternative to the \code{repel} option.}

\item{repel}{If set to \code{TRUE}, will try to reposition the text label to avoid overlapping it with other elements of the plot (such as the lines). It relies on the \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}} function of the \link{ggrepel} package.}

\item{box.padding}{Amount of padding around bounding box, as unit or number.
Defaults to 0.25. (Default unit is lines, but other units can be specified
by passing \code{unit(x, "units")}).}

\item{point.padding}{Amount of padding around labeled point, as unit or
number. Defaults to 0. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{min.segment.length}{Skip drawing segments shorter than this, as unit or
number. Defaults to 0.5. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{arrow}{specification for arrow heads, as created by \code{\link[grid]{arrow}}}

\item{force}{Force of repulsion between overlapping text labels. Defaults
to 1.}

\item{force_pull}{Force of attraction between a text label and its
corresponding data point. Defaults to 1.}

\item{max.time}{Maximum number of seconds to try to resolve overlaps.
Defaults to 0.5.}

\item{max.iter}{Maximum number of iterations to try to resolve overlaps.
Defaults to 10000.}

\item{max.overlaps}{Exclude text labels that overlap too many things.
Defaults to 10.}

\item{xlim, ylim}{Limits for the x and y axes. Text labels will be constrained
to these limits. By default, text labels are constrained to the entire plot
area.}

\item{direction}{"both", "x", or "y" -- direction in which to adjust position of labels}

\item{seed}{Random seed passed to \code{\link[base]{set.seed}}. Defaults to
\code{NA}, which means that \code{set.seed} will not be called.}

\item{verbose}{If \code{TRUE}, some diagnostics of the repel algorithm are printed}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}} or
\code{\link[ggplot2]{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{text_aes}{Named list, additional aesthetics to send to the geometry.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. There are
  three types of arguments you can use here:

  \itemize{
    \item Aesthetics: to set an aesthetic to a fixed value, like
       \code{colour = "red"} or \code{size = 3}.
    \item Other arguments to the layer, for example you override the
      default \code{stat} associated with the layer.
    \item Other arguments passed on to the stat.
  }}
}
\description{
This function annotates the chart with the last date where data is available.
It does so by looking by first filtering out missing values in the y-axis
variable and then computing for these rows the maximum of the x-axis variable.
}
\details{
This function formats the last date string accordingly to its frequency
and to publication standards. For instance, monthly data gets formatted
as "Jan.", "Feb.", etc. with the exception of June and September that
appear as "June" and "Sept." respectively.

Note that this formatting is currently incomplete for frequencies other
than monthly.
}
