% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{annotate_top_tick}
\alias{annotate_top_tick}
\title{Add wide ticks at the top of the chart}
\usage{
annotate_top_tick()
}
\description{
This function adds ticks to the top-left and top-right corners of the charts.
These ticks are twice as long as normal ticks and serve mostly as as
a decoration.
}
