% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{bsvr_colors}
\alias{bsvr_colors}
\title{List of colors used in BSVR reports}
\format{
An object of class \code{character} of length 9.
}
\usage{
bsvr_colors
}
\description{
List of colors used in BSVR reports
}
\keyword{datasets}
