% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession_bars.R
\docType{data}
\name{customproto}
\alias{customproto}
\alias{GeomRecessions}
\title{Custom ggproto classes}
\description{
The \code{fedplot} package contains a few custom ggproto objects. For the
most part, these are slightly tweaked versions of ggplot2's default proto
objects. For more information about these, see \link[ggplot2:ggproto]{ggplot2::ggproto}.
}
\section{Functions}{
\itemize{
\item \code{GeomRecessions}: Add recession bars.

}}
\keyword{datasets}
