% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{.bigpt}
\alias{.bigpt}
\title{Alternative to ggplot::.pt ; using bigpoints of size 72 instead of points of size 72.27.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.bigpt
}
\description{
Alternative to ggplot::.pt ; using bigpoints of size 72 instead of points of size 72.27.
}
\keyword{internal}
