% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{.pt}
\alias{.pt}
\title{ggpplot2 fed theme, including advanced tweaks}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.pt
}
\description{
ggpplot2 fed theme, including advanced tweaks
}
\keyword{datasets}
