% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{draw_key_square}
\alias{draw_key_square}
\title{Square legends}
\usage{
draw_key_square(data, params, size)
}
\description{
By \href{https://github.com/tidyverse/ggplot2/issues/3669}{default}, legends of bar plots and area plots use the \code{draw_key_polygon} function
to draw a rectangle as high as the text height. \code{fedplot} instead uses square legends, achieved by passing the argument \code{key_glyph=draw_key_square}
to the required geometries (\code{geom_col}, etc.).
See this \href{https://stackoverflow.com/questions/65812949/set-standard-legend-key-size-with-long-label-names-ggplot}{stack overflow thread} for more details,
and \href{https://www.emilhvitfeldt.com/post/changing-glyph-in-ggplot2/}{this page} for examples of other legend key functions.
}
