% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{dummy_data}
\alias{dummy_data}
\title{Internal ggplot2 functions that we have to copy (CRAN doesn't support :::)}
\usage{
dummy_data()
}
\description{
Internal ggplot2 functions that we have to copy (CRAN doesn't support :::)
}
\keyword{internal}
