% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fed_color_pal}
\alias{fed_color_pal}
\title{Return function to interpolate color palette}
\usage{
fed_color_pal(palette = "fsr_primary", use_ramp = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in fed_palettes}

\item{use_ramp}{Boolean indicating whether to use a color ramp to interpolate colors (via \link[grDevices:colorRamp]{grDevices::colorRampPalette})}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to \code{colorRampPalette()}}
}
\description{
Return function to interpolate color palette
}
