% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fed_linewidth_pal}
\alias{fed_linewidth_pal}
\title{Closure to generate a linewidth palette picker, given an input palette}
\usage{
fed_linewidth_pal(palette = fsr_linewidths)
}
\arguments{
\item{palette}{Vector of linewidths that will be used as a palette; must be larger than the maximum number of lines plotted.}
}
\value{
function
}
\description{
This internal function will take as an input a palette (a vector of numbers
representing linewidths) and return a function \code{f(n)} that produces a
smaller vector containing the first \code{n} linewidths of the
input palette.
}
