% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedplot-package.R
\docType{package}
\name{fedplot}
\alias{fedplot}
\alias{fedplot-package}
\title{fedplot}
\description{
Fed-style \pkg{ggplot2} theme and functions.
}
\details{
This package contains a collection of ggplot extensions,
(a theme, color palettes, and geoms) required to produce
Fed-style charts, similar to those produced for the FSR.
Detailed documentation can be viewed at
\url{https://sergiocorreia.github.io/fedplot}.

Please report issues and suggest improvements at
\url{https://github.com/sergio.correia/fedplot/issues}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sergiocorreia.github.io/fedplot}
  \item \url{https://github.com/sergiocorreia/fedplot}
  \item Report bugs at \url{https://github.com/sergiocorreia/fedplot/issues}
}

}
\author{
\strong{Maintainer}: Sergio Correia \email{sergio.a.correia@frb.gov} (\href{https://orcid.org/0000-0002-0914-8648}{ORCID})

}
\keyword{internal}
