% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{fontsupport}
\alias{fontsupport}
\title{Font support}
\description{
\code{fedplot} was designed with a specific set of fonts in mind.
You can use the \code{\link[=install_fed_font]{install_fed_font()}} and
\code{\link[=load_fed_font]{load_fed_font()}} to install a free version of these fonts.
}
\details{
Note that these functions are currently only partly implemented across
operating systems. For instance, \code{install_fed_font} is only implemented
for Linux, and \code{load_fed_font} for Windows.

Lastly, note you can also override the font used through \code{\link[=options]{options()}}, such as in:
}
