% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FSR_1_17}
\alias{FSR_1_17}
\title{Change in house prices}
\format{
A data frame with 711 rows and 3 columns, showing the 12-month change
in the house price index from three different sources.
\describe{
\item{date}{Month}
\item{source}{Data provider}
\item{value}{12-month percent change in the house price index}
}
}
\source{
\url{https://www.federalreserve.gov/publications/files/financial-stability-report-20221104.pdf}
}
\usage{
FSR_1_17
}
\description{
Sample dataset from Figure 1.17 of the November 2022 Financial Stability Report,
used to illustrate the use of the \code{fedplot} package.
This file can be used to illustrate line plots.
}
\keyword{datasets}
