% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FSR_4_1}
\alias{FSR_4_1}
\title{Runnable money-like liabilities (\% of GDP)}
\format{
A data frame with 83 rows and 3 columns, showing the
amount of runnable money-like liabilities across time,
as a share of GDP, by type of liability.
\describe{
\item{date}{Quarter}
\item{type}{Type of runnable money-like liability (six categories:
Uninsured deposits, Repurchase agreements, Domestic money market funds, Commercial paper, Securities lending, and Other.)}
\item{value}{Net issuance of a given type of risky debt, in $bn}
}
}
\source{
\url{https://www.federalreserve.gov/publications/files/financial-stability-report-20221104.pdf}
}
\usage{
FSR_4_1
}
\description{
Sample dataset from Figure 4.1 of the November 2022 Financial Stability Report,
used to illustrate the use of the \code{fedplot} package.
This file can be used to illustrate area plots.
}
\keyword{datasets}
