% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{fsr_linewidths}
\alias{fsr_linewidths}
\title{Linewidths used in the FSR.}
\format{
An object of class \code{numeric} of length 14.
}
\usage{
fsr_linewidths
}
\description{
Linewidths used in the FSR.
}
\keyword{datasets}
