% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{fsr_primary_colors}
\alias{fsr_primary_colors}
\title{List of primary colors used in the BSVR}
\format{
An object of class \code{character} of length 4.
}
\usage{
fsr_primary_colors
}
\description{
List of primary colors used in the BSVR
}
\keyword{datasets}
