% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{fsr_secondary_colors}
\alias{fsr_secondary_colors}
\title{List of secondary colors used in the BSVR}
\format{
An object of class \code{character} of length 4.
}
\usage{
fsr_secondary_colors
}
\description{
List of secondary colors used in the BSVR
}
\keyword{datasets}
