% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{geom_hline_zero}
\alias{geom_hline_zero}
\title{Convenience functions that wrap existing ggplot2 code.
This file shouldn't contain new functionality, just code
that can be replicated directly with ggplot2 standard functions.
Draw horizontal line at the origin (y=0)
Equivalent to geom_hline(yintercept = 0, linewidth = getOption("fedplot.linewidth_adj"), lineend = "round")}
\usage{
geom_hline_zero(...)
}
\arguments{
\item{...}{Additional arguments to pass to \link[ggplot2:geom_abline]{ggplot2::geom_hline}}
}
\description{
Convenience functions that wrap existing ggplot2 code.
This file shouldn't contain new functionality, just code
that can be replicated directly with ggplot2 standard functions.
Draw horizontal line at the origin (y=0)
Equivalent to geom_hline(yintercept = 0, linewidth = getOption("fedplot.linewidth_adj"), lineend = "round")
}
