% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{geom_line_fed}
\alias{geom_line_fed}
\title{Draw lines with Fed-style aesthetics}
\usage{
geom_line_fed(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Additional \code{aes()} arguments passed to \link[ggplot2:geom_path]{ggplot2::geom_line}}

\item{...}{Additional arguments passed to \link[ggplot2:geom_path]{ggplot2::geom_line}}
}
\description{
\code{geom_line_fed} is a wrapper around \code{ggplot2::geom_line()}
that adds the required line colors and line widths.
It also sets the \code{linejoin} and \code{lineend} attributes to their required values.
}
\details{
It is equivalent to:

\if{html}{\out{<div class="sourceCode">}}\preformatted{geom_line(aes(color=<VAR>, linewidth=<VAR>),
          na.rm = TRUE,
          linejoin = "mitre",
          lineend = "round")
}\if{html}{\out{</div>}}
}
