% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession_bars.R
\name{geom_recessions}
\alias{geom_recessions}
\title{Add recession bars to time series graphs}
\usage{
geom_recessions(
  fill = "#BDCFDE",
  alpha = 1,
  draw_top_bar = TRUE,
  top_fill = "#236192",
  top_alpha = 1
)
}
\arguments{
\item{fill}{The color of the main bar; defaults to a light blue tint.}

\item{alpha}{The alpha transparency of the main bar; defaults to 1.0 (totally opaque).}

\item{draw_top_bar}{Whether to draw the top bar or not.}

\item{top_fill}{The color of the top bar; defaults to a dark blue tint.}

\item{top_alpha}{The alpha transparency of the top blue bar; defaults to 1.0 (totally opaque).}
}
\description{
\code{geom_recessions} adds shaded recession bars to a time-series ggplot2 chart
(a chart with a date object in the x-axis.) These recession bars correspond to all
\href{https://www.nber.org/research/data/us-business-cycle-expansions-and-contractions}{NBER-dated} recessions
that overlap with the date range of the chart.
By default, each recession bar will be drawn as a light blue rectangle
spanning the entirety of the y-axis plus a dark blue rectangle drawn only over
the topmost part of the chart.
}
\section{Implementation details}{

Internally, \code{geom_recessions} works by adding a \link{GeomRecessions}
\link[ggplot2:ggproto]{ggplot2::ggproto} object, which inherits from \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}. It
replaces the \code{setup_data} and \code{draw_panel} methods in order to
load the recession data and draw the bars respectively.
Further, the actual drawing is done by two \link[grid:grid.rect]{grid::rectGrob}
objects, for the main blue bar and the top dark blue bar.
}

\section{Inspiration and alternative implementations}{

This function is heavily inspired on \code{geom_recessions} from the \href{https://cmap-repos.github.io/cmapplot/reference/geom_recessions.html}{CMAPPLOT} package.
}

