% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\docType{data}
\name{GeomLastDate}
\alias{GeomLastDate}
\title{ggproto class used to place the text by annotate_last_date, based on GeomText}
\format{
An object of class \code{GeomLastDate} (inherits from \code{GeomText}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\description{
ggproto class used to place the text by annotate_last_date, based on GeomText
}
\seealso{
ggproto
}
\keyword{datasets}
