% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\docType{data}
\name{GeomLastDateRepel}
\alias{GeomLastDateRepel}
\title{ggproto class used to place the text by annotate_last_date, based on GeomTextRepel}
\format{
An object of class \code{GeomLastDateRepel} (inherits from \code{GeomTextRepel}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\description{
ggproto class used to place the text by annotate_last_date, based on GeomTextRepel
}
\seealso{
ggproto
}
\keyword{datasets}
