% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedplot-package.R
\docType{import}
\name{ggplot_add}
\alias{ggplot_add}
\title{Add custom objects to ggplot}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{ggplot_add}}}
}}

