% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedplot-package.R
\name{ggplot_addx}
\alias{ggplot_addx}
\alias{ggplot__add}
\title{Add custom objects to ggplot}
\usage{
ggplot__add(object, plot, object_name)
}
\description{
Add custom objects to ggplot
}
\keyword{internal}
