% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_fed_font}
\alias{install_fed_font}
\title{Installing fonts}
\usage{
install_fed_font()
}
\value{
None
}
\description{
Installing fonts is a tricky and system-dependent process.
Currently, only Linux font install is supported and Windows fonts should
installed manually.
}
\details{
Notes:
\itemize{
\item Do not use \code{showtext}; it is not useful as it does not embed the fonts into the pdf/eps.
\item \code{extrafont} had some issues on Linux.
\item TODO: maybe we can just load from \code{extdata} without installing into \verb{~/.fonts}?
}
}
