% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{load_fed_font}
\alias{load_fed_font}
\title{Loading fonts}
\usage{
load_fed_font()
}
\value{
None
}
\description{
Load an already-installed font into R
}
\details{
Currently only available for Windows, in which case it uses \code{extrafont::loadfonts()}.
}
