% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_plot}
\alias{save_plot}
\title{Export figure}
\usage{
save_plot(
  filename,
  extension = c("pdf", "eps", "png", "all"),
  path = NULL,
  scale = 1,
  dpi = 300,
  plot = ggplot2::last_plot(),
  ...
)
}
\arguments{
\item{filename}{filename to be saved, excluding the path and extension.}

\item{extension}{image extension; also defines the type of image. Valid values are c('pdf', 'eps', 'png', 'all'); note that 'all' will create all image types.}

\item{path}{path where the file(s) will be saved}

\item{scale}{rescaling done by \link[ggplot2:ggsave]{ggplot2::ggsave}; experimental.}

\item{dpi}{dots per inch of the output file(s). Typical values are 300 for print (the default) and rarely 72 for web image output. Note that for png images \code{save_plot()} multiples the dpi by 2, to compensate for the lower quality compared to vector formats (pdf, eps). Thus, the default dpi for png images is 600.}

\item{plot}{ggplot2 to save. By default will save the plot last printed to screen.}

\item{...}{Other arguments passed on to the graphics device function.}
}
\description{
Export figure
}
