% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{set_plot_size}
\alias{set_plot_size}
\title{Set plot size}
\usage{
set_plot_size(
  size = c("narrow", "wide", "box_narrow", "box_wide", "slides", "custom"),
  plot_height = NULL,
  plot_width = NULL,
  panel_height = NULL,
  panel_width = NULL
)
}
\arguments{
\item{size}{dimensions of the figure. For instance, 'narrow' creates default figures of 192x99 points. Valid values are c('narrow', 'wide', 'box_narrow', 'box_wide', 'slides', 'custom'). Note that 'slides' will produce a 192x161 figure, ideal for two-pane Powerpoint slides.}

\item{plot_height}{height of the figure in points (technically; bigpoints); useful when size='custom'.}

\item{plot_width}{height of the figure in points (technically; bigpoints); useful when size='custom'.}

\item{panel_height}{height of the panel in points (technically; bigpoints); useful when size='custom'.}

\item{panel_width}{height of the panel in points (technically; bigpoints); useful when size='custom'.}
}
\value{
A \code{ggh4x::forcedsize} S3 object that can be added to a plot.
}
\description{
This function sets the size of the plot, in big points.
By default, it will create a figure with width equal to 192bp and
panel height (not total figure height!) equal to 99bp.
It supports other present sizes as well as custom sizes.
}
