% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fed_color_pal}
\alias{fed_color_pal}
\title{Closure to generate a color palette picker, given an input palette}
\usage{
fed_color_pal(palette = "fsr_primary", use_ramp = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in fed_color_palettes}

\item{use_ramp}{Boolean indicating whether to use a color ramp to interpolate colors (via \link[grDevices:colorRamp]{grDevices::colorRampPalette})}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to \code{colorRampPalette()}}
}
\value{
function
}
\description{
This internal function will take as an input either a) the name of a palete (
one of \code{fsr_primary_colors}, \code{fsr_secondary_colors}, or \code{bsvr_color})
or b) the palette itself as a vector of hex codes representing the colors.
In turn, it will return a function \code{f(n)} that produces a vector
containing the first \code{n} colors of the input palette.
Optionally, it can use the input palette to generate a color ramp, so
it can produce as many colors as required.
}
