% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{move_y_axis_title}
\alias{move_y_axis_title}
\title{Complex extensions that involve hacks to ggplot internals}
\usage{
move_y_axis_title()
}
\description{
Complex extensions that involve hacks to ggplot internals
}
