% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_theme.R
\name{theme_fed_minimal}
\alias{theme_fed_minimal}
\title{Internal Fed theme}
\usage{
theme_fed_minimal(
  font_family = getOption("fedplot.font_family"),
  font_size = getOption("fedplot.font_size"),
  legend_position = c(0.5, 0.9)
)
}
\arguments{
\item{font_family}{base font family}

\item{font_size}{base font family and size}

\item{legend_position}{where to place the legend; a tuple indicating the top- and left-most corners of the legend. Default is \code{c(.5, .9)}.}
}
\description{
The \code{theme_fed()} object is composed of a theme, plus a number of functions
that achieve features that the theme itself cannot achieve.
This function corresponds to the theme itself, modifying \code{ggplot2::theme_classic}.
}
